pyproject.toml
[tool.poetry]
name = "web3_utilities"
version = "0.1.0"
description = "Various utilities and helper methods for working with Web3, primarily utilizing Python"
authors = ["Michael Trosen <mrtrosen@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.9"
web3 = "^5.29.2"
arrow = "^1.2.2"
python-decouple = "^3.6"
rich = "^12.4.4"
click = "^8.1.3"

[tool.poetry.dev-dependencies]
black = "^22.3.0"
isort = "^5.10.1"
pre-commit = "^2.19.0"
autoflake = "^1.4"
flake8 = "^4.0.1"
autopep8 = "^1.6.0"
safety = "^1.10.3"
pytest = "^7.1.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 140
target-version = ['py39']
include = '(src\/scripts\/.*$|\.pyi?$)'
exclude = '''
/(
    \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''

[tool.pylint.format]
max-line-length = "140"

[tool.isort]
profile = "black"
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
line_length = 140
